# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unified Trading API",
    "description": "Consolidated API gateway \u2014 61 endpoints across 16 domains",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analytics/pnl": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Pnl",
        "description": "Get PnL breakdown.",
        "operationId": "get_pnl_analytics_pnl_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "venue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Venue"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1d",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Pnl Analytics Pnl Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Create Pnl Snapshot",
        "description": "Trigger a PnL snapshot calculation.",
        "operationId": "create_pnl_snapshot_analytics_pnl_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Pnl Snapshot Analytics Pnl Post"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/timeseries": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Timeseries",
        "description": "Get analytics timeseries data.",
        "operationId": "get_timeseries_analytics_timeseries_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "equity",
              "title": "Metric"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1d",
              "title": "Period"
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1h",
              "title": "Granularity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Timeseries Analytics Timeseries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Create Timeseries Entry",
        "description": "Add a timeseries data point.",
        "operationId": "create_timeseries_entry_analytics_timeseries_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Timeseries Entry Analytics Timeseries Post"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/performance": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Performance",
        "description": "Get performance metrics.",
        "operationId": "get_performance_analytics_performance_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "30d",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Performance Analytics Performance Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Create Performance Snapshot",
        "description": "Trigger a performance snapshot.",
        "operationId": "create_performance_snapshot_analytics_performance_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Performance Snapshot Analytics Performance Post"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/organizations": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Organizations",
        "description": "Get analytics organizations.",
        "operationId": "get_organizations_analytics_organizations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Organizations Analytics Organizations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analytics/settlements": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Settlements",
        "description": "Get settlement records.",
        "operationId": "get_settlements_analytics_settlements_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Settlements Analytics Settlements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Create Settlement",
        "description": "Create a settlement record.",
        "operationId": "create_settlement_analytics_settlements_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Settlement Analytics Settlements Post"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/instruments": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Instruments",
        "description": "Get analytics instrument list.",
        "operationId": "get_instruments_analytics_instruments_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "asset_class",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Asset Class"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Instruments Analytics Instruments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/organizations": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Organizations",
        "description": "Get organizations.",
        "operationId": "get_organizations_users_organizations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Organizations Users Organizations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/users/members": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Members",
        "description": "Get organization members.",
        "operationId": "get_members_users_members_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Organization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Members Users Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/subscriptions": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Subscriptions",
        "description": "Get subscription plans and status.",
        "operationId": "get_subscriptions_users_subscriptions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Subscriptions Users Subscriptions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}